Load sepolicy rules from every available partitions
Some devices has /cache or /metadata in their fstab but doesn't actually mount it. We manually find the directory to store sepolicy rules if the directory was given by magiskinit is not available, and magiskinit now will load sepolicy rules from every available partitions to support this case.
Close topjohnwu#4642